requirements.txt
cffi==1.9.1
click==6.6
cryptography==1.7.2
decorator==4.0.10
Django==1.10.5
django-braces==1.11.0
django-filter==0.14.0
django-oauth-toolkit==0.11.0
django-rest-framework-social-oauth2==1.0.5
django-storages-redux==1.3.2
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
enum34==1.1.6
geocoder==1.19.0
googlemaps==2.4.5
gunicorn==19.6.0
idna==2.2
ipaddress==1.0.18
Markdown==2.6.6
ndg-httpsclient==0.4.2
numpy==1.11.3
oauthlib==1.1.2
pandas==0.19.2
Pillow==3.3.1
psycopg2==2.6.2
pusher==1.6
pyasn1==0.2.2
pycparser==2.17
pyfcm==1.0.7
PyJWT==1.4.2
pyOpenSSL==16.2.0
python-dateutil==2.6.0
python-magic==0.4.12
python-openid==2.2.5
pytz==2016.10
ratelim==0.1.6
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
urllib3==1.17
validators==0.11.1
